Restructured functional tests
