+++ tests/inspiretests/Makefile
