+++ tests/inspiretests/inspire.cpp
